refactor: standardize determined info store
