+++ webui/react/src/ioTypes.ts
@@ -25,25 +25,9 @@ export const decode = <T>(type: io.Mixed, data: any): T => {
   }
 };
 
-const optional = (x: io.Mixed) => io.union([x, io.null, io.undefined]);
-
-/* Info */
-
-const ioSsoProvider = io.type({
-  name: io.string,
-  sso_url: io.string,
-});
-
-export const ioDeterminedInfo = io.type({
-  cluster_id: io.string,
-  cluster_name: io.string,
-  isTelemetryEnabled: io.boolean,
-  master_id: io.string,
-  sso_providers: optional(io.array(ioSsoProvider)),
-  version: io.string,
-});
-
-export type ioTypeDeterminedInfo = io.TypeOf<typeof ioDeterminedInfo>;
+export const optional = (x: io.Mixed): io.Mixed | io.NullC | io.UndefinedC => {
+  return io.union([x, io.null, io.undefined]);
+};
 
 /* Slot */
